Moved inheritance-dependent HighscoreList-code to a different package
