+++ no.hal.jex.collection/src/patterns/observable/inheritance/HighscoreProgram.java
@@ -1,4 +1,4 @@
-package patterns.observable;
+package patterns.observable.inheritance;
 
 import java.util.Scanner;
