+++ no.hal.jex.collection/src/patterns/observable/inheritance/ListListener.java
@@ -1,4 +1,4 @@
-package patterns.observable;
+package patterns.observable.inheritance;
 
 public interface ListListener {
 	public void listChanged(ObservableList list, int pos);
